Revert "Code update"
This reverts commit ae9d6fd90f9d1e557284cefb6dc7c12ff8e6fac8.

+++ cu.bat
@@ -6,5 +6,6 @@ for /D %%d in (*) do (
    cd ..                                           
 )                                                  
 
-del /s /q *.orig 
-del /s /q *.o 
+if exist *.orig del /s /q *.orig 
+if exist *.b32 del /s /q *.b32 
+if exist *.o del /s /q *.o
